requirements.txt
django>=2.2.5,<2.3
django-cleanup>=4.0.0,<5.0
django-extensions>=2.2.1,<2.3
django-jsonview>=1.2.0,<2.0
django-redis>=4.10,<5.0
djangorestframework>=3.10.3,<4.0
gunicorn>=19.9.0,<20.0
lxml>=4.4.1,<5.0
psycopg2-binary==2.8.3
pydub==0.23.1
python-magic>=0.4.15,<0.5
redis>=3.3.11,<4.0
requests>=2.22,<3.0
sox==1.3.7
typing-extensions>=3.7.4,<4.0
unidecode>=1.1.1,<1.2
xlrd>=1.2,<2.0
xlwt>=1.3,<2.0
